+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildTarMojo.java
@@ -26,6 +26,7 @@ import com.google.cloud.tools.jib.plugins.common.BuildStepsRunner;
 import com.google.cloud.tools.jib.plugins.common.HelpfulSuggestions;
 import com.google.cloud.tools.jib.plugins.common.InferredAuthRetrievalException;
 import com.google.cloud.tools.jib.plugins.common.InvalidAppRootException;
+import com.google.cloud.tools.jib.plugins.common.InvalidContainerVolumeException;
 import com.google.cloud.tools.jib.plugins.common.InvalidWorkingDirectoryException;
 import com.google.cloud.tools.jib.plugins.common.MainClassInferenceException;
 import com.google.cloud.tools.jib.plugins.common.PluginConfigurationProcessor;
@@ -115,6 +116,9 @@ public class BuildTarMojo extends JibPluginConfiguration {
               + ex.getInvalidPathValue(),
           ex);
 
+    } catch (InvalidContainerVolumeException ex) {
+      throw new MojoExecutionException(
+          "<container><volumes> is not an absolute Unix-style path: " + ex.getInvalidVolume(), ex);
     } catch (InvalidImageReferenceException
         | IOException
         | CacheDirectoryCreationException