+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenRawConfiguration.java
@@ -133,6 +133,11 @@ class MavenRawConfiguration implements RawConfiguration {
   }
 
   @Override
+  public List<String> getVolumes() {
+    return jibPluginConfiguration.getVolumes();
+  }
+
+  @Override
   public List<String> getPorts() {
     return jibPluginConfiguration.getExposedPorts();
   }