+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/PluginConfigurationProcessor.java
@@ -41,6 +41,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.time.Instant;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -131,7 +132,8 @@ public class PluginConfigurationProcessor {
       @Nullable Map<String, String> dockerEnvironment,
       HelpfulSuggestions helpfulSuggestions)
       throws InvalidImageReferenceException, MainClassInferenceException, InvalidAppRootException,
-          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException {
+          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException,
+          InvalidContainerVolumeException {
     ImageReference targetImageReference =
         getGeneratedTargetDockerTag(rawConfiguration, projectProperties, helpfulSuggestions);
     DockerDaemonImage targetImage = DockerDaemonImage.named(targetImageReference);
@@ -153,7 +155,8 @@ public class PluginConfigurationProcessor {
       Path tarImagePath,
       HelpfulSuggestions helpfulSuggestions)
       throws InvalidImageReferenceException, MainClassInferenceException, InvalidAppRootException,
-          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException {
+          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException,
+          InvalidContainerVolumeException {
     ImageReference targetImageReference =
         getGeneratedTargetDockerTag(rawConfiguration, projectProperties, helpfulSuggestions);
     TarImage targetImage = TarImage.named(targetImageReference).saveTo(tarImagePath);
@@ -167,7 +170,7 @@ public class PluginConfigurationProcessor {
       RawConfiguration rawConfiguration, ProjectProperties projectProperties)
       throws InferredAuthRetrievalException, InvalidImageReferenceException,
           MainClassInferenceException, InvalidAppRootException, IOException,
-          InvalidWorkingDirectoryException {
+          InvalidWorkingDirectoryException, InvalidContainerVolumeException {
     Preconditions.checkArgument(rawConfiguration.getToImage().isPresent());
 
     ImageReference targetImageReference = ImageReference.parse(rawConfiguration.getToImage().get());
@@ -209,7 +212,8 @@ public class PluginConfigurationProcessor {
       ImageReference targetImageReference,
       boolean isTargetImageCredentialPresent)
       throws InvalidImageReferenceException, MainClassInferenceException, InvalidAppRootException,
-          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException {
+          InferredAuthRetrievalException, IOException, InvalidWorkingDirectoryException,
+          InvalidContainerVolumeException {
     JibSystemProperties.checkHttpTimeoutProperty();
 
     ImageReference baseImageReference =
@@ -247,6 +251,7 @@ public class PluginConfigurationProcessor {
             .setProgramArguments(rawConfiguration.getProgramArguments().orElse(null))
             .setEnvironment(rawConfiguration.getEnvironment())
             .setExposedPorts(ExposedPortsParser.parse(rawConfiguration.getPorts()))
+            .setVolumes(getVolumesList(rawConfiguration))
             .setLabels(rawConfiguration.getLabels())
             .setUser(rawConfiguration.getUser().orElse(null));
     getWorkingDirectoryChecked(rawConfiguration)
@@ -268,6 +273,29 @@ public class PluginConfigurationProcessor {
         targetImageReference,
         isBaseImageCredentialPresent,
         isTargetImageCredentialPresent);
+  }
+
+  /**
+   * Parses the list of raw volumes directories to a list of {@link AbsoluteUnixPath}
+   *
+   * @param rawConfiguration raw configuration data
+   * @return the list of parsed volumes.
+   * @throws InvalidContainerVolumeException if {@code volumes} are not valid absolute Unix paths
+   */
+  @VisibleForTesting
+  static List<AbsoluteUnixPath> getVolumesList(RawConfiguration rawConfiguration)
+      throws InvalidContainerVolumeException {
+    List<AbsoluteUnixPath> volumes = new ArrayList<>();
+    for (String path : rawConfiguration.getVolumes()) {
+      try {
+        AbsoluteUnixPath absoluteUnixPath = AbsoluteUnixPath.get(path);
+        volumes.add(absoluteUnixPath);
+      } catch (IllegalArgumentException exception) {
+        throw new InvalidContainerVolumeException(path, path, exception);
+      }
+    }
+
+    return volumes;
   }
 
   /**